pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "npu-model"
version = "0.1.0"
description = "Plugin-based model converter producing accelerator-targeted ORT GenAI bundles."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "dsmith111" }]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "huggingface_hub>=0.23.0",
  "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "types-tqdm",
]
export = [
  "onnxruntime-genai",
  "torch",
  "transformers",
  "onnx",
  "onnx-ir",
]
qnn = [
  "onnxruntime-qnn",
]
validate = [
  "onnxruntime-genai",
]

[project.scripts]
npu-model = "npu_model.cli.main:app"

# Plugin entry points (built-ins in this repo)
[project.entry-points."npu_model.backends"]
qnn = "npu_model.backends.qnn:QnnBackend"

[project.entry-points."npu_model.adapters"]
phi3 = "npu_model.adapters.phi3:Phi3Adapter"
llama = "npu_model.adapters.llama:LlamaAdapter"

[project.entry-points."npu_model.runtime_formats"]
ort-genai-folder = "npu_model.runtime_formats.ort_genai_folder:OrtGenaiFolderFormat"

[project.entry-points."npu_model.quantizers"]
passthrough = "npu_model.quant.passthrough:PassthroughQuantizer"
qnn-qdq = "npu_model.quant.qnn_qdq:QnnQdqQuantizer"

[project.entry-points."npu_model.exporters"]
ort-genai-builder = "npu_model.exporters.ort_genai_builder:OrtGenaiBuilderExporter"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 110
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
no_implicit_optional = true
